+++ CHANGELOG.md
@@ -6,7 +6,7 @@ This file contains the list of changes made to the Fitterbap library.
 
 ## 0.4.0
 
-2021 Jun 3
+2021 Jun 4
 
 *   Improved comm stack based upon EOC2021 attendee feedback - thank you!
     *   Added framer length crc-8 and mandatory EOF match.
@@ -25,6 +25,7 @@ This file contains the list of changes made to the Fitterbap library.
     Returns INT64_MAX rather than -1 on no scheduled events to simplify caller.
 *   Added fbp_evm_on_schedule for improved event_manager thread integration. 
 *   Fixed deadlock between GIL and native mutex.
+*   Added log message port to comm stack.
 
 
 ## 0.3.2